pyproject.toml
[project]
name = "pixelblaze-mcp"
version = "0.1.0"
description = "MCP server for AI-powered PixelBlaze LED controller interaction"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.4.0",
    "pixelblaze-client>=1.1.7",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pixelblaze-mcp = "pixelblaze_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pixelblaze_mcp"]
